add error handler

+++ src/main/java/com/dosmakhambetbaktiyar/module41/service/impl/CarDataServiceImpl.java
@@ -2,12 +2,8 @@ package com.dosmakhambetbaktiyar.module41.service.impl;
 
 import com.dosmakhambetbaktiyar.module41.model.CartData;
 import com.dosmakhambetbaktiyar.module41.repository.CartDataRepository;
-import com.dosmakhambetbaktiyar.module41.security.CustomPrincipal;
 import com.dosmakhambetbaktiyar.module41.service.CartDataService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.ReactiveSecurityContextHolder;
-import org.springframework.security.core.context.SecurityContext;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
@@ -18,6 +14,10 @@ public class CarDataServiceImpl implements CartDataService {
     private  CartDataRepository repository;
     @Override
     public Mono<CartData> save(CartData cartData) {
+        if(cartData.getId() != null){
+            return Mono.error(new RuntimeException("Cart data id should be null for saving new cart data."));
+        }
+
         return repository.save(cartData)
                 .onErrorMap(e -> new RuntimeException("Error occurred while saving cart data " + e.getMessage()));
     }
@@ -31,27 +31,15 @@ public class CarDataServiceImpl implements CartDataService {
 
     @Override
     public Flux<CartData> findAll() {
-
         return repository.findAll();
     }
 
     @Override
-    public Mono<String> findAll2() {
-
-        return ReactiveSecurityContextHolder.getContext()
-                .map(SecurityContext::getAuthentication)
-                .map(Authentication::getPrincipal)
-                .cast(CustomPrincipal.class) // Replace with your principal class
-                .map(CustomPrincipal::getName) // Replace with your method to get user id
-                .flatMap(userId -> {
-                    // Use userId here
-                    System.out.println("User ID: " + userId);
-                    return Mono.just(userId);
-                });
-    }
-
-    @Override
     public Mono<CartData> update(CartData cartData) {
+        if(cartData.getId() == null){
+            return Mono.error(new RuntimeException("Cart data id should not be null for updating cart data."));
+        }
+
         return repository.save(cartData)
                 .onErrorMap(e -> new RuntimeException("Error occurred while updating cart data"));
     }
@@ -64,11 +52,10 @@ public class CarDataServiceImpl implements CartDataService {
 
     @Override
     public void delete(CartData cartData) {
-        repository.delete(cartData);
-    }
+        if(cartData.getId() == null){
+            throw new RuntimeException("Cart data id should not be null for deleting cart data.");
+        }
 
-    @Override
-    public Mono<CartData> findByCartNumber(String cartNumber) {
-        return repository.findByCartNumber(cartNumber);
+        repository.delete(cartData);
     }
 }
